refactor(copilot): move widget modules into lib
