+++ apps/tradinggoose/lib/copilot/components/copilot-app.test.tsx
