+++ apps/tradinggoose/lib/copilot/components/copilot-message/components/assistant-message-segments.test.ts
