+++ apps/tradinggoose/lib/copilot/components/copilot-message/components/file-display.tsx
