+++ apps/tradinggoose/lib/copilot/components/copilot-message/components/index.ts
