+++ apps/tradinggoose/lib/copilot/components/copilot-message/components/markdown-renderer.tsx
