+++ apps/tradinggoose/lib/copilot/components/copilot-message/components/options-selector.tsx
