+++ apps/tradinggoose/lib/copilot/components/copilot-message/components/smooth-streaming.tsx
