+++ apps/tradinggoose/lib/copilot/components/copilot-message/components/thinking-group.test.tsx
