+++ apps/tradinggoose/lib/copilot/components/copilot-message/copilot-message.test.tsx
