+++ apps/tradinggoose/lib/copilot/components/copilot-message/message-visibility.test.ts
