+++ apps/tradinggoose/lib/copilot/components/index.ts
