+++ apps/tradinggoose/lib/copilot/components/todo-list/todo-list.tsx
