+++ apps/tradinggoose/lib/copilot/components/user-input/components/access-level-selector.tsx
