+++ apps/tradinggoose/lib/copilot/components/user-input/components/attached-files-grid.tsx
