+++ apps/tradinggoose/lib/copilot/components/user-input/components/mention-menu.tsx
