+++ apps/tradinggoose/lib/copilot/components/user-input/components/model-selector.tsx
