+++ apps/tradinggoose/lib/copilot/components/user-input/constants.ts
