+++ apps/tradinggoose/lib/copilot/components/user-input/hooks/use-user-input-attachments.ts
