+++ apps/tradinggoose/lib/copilot/components/user-input/hooks/use-user-input-mention-sources.ts
