+++ apps/tradinggoose/lib/copilot/components/user-input/mention-copy.ts
