+++ apps/tradinggoose/lib/copilot/components/user-input/mention-editor-dom.test.ts
