+++ apps/tradinggoose/lib/copilot/components/user-input/mention-utils.ts
