+++ apps/tradinggoose/lib/copilot/components/user-input/types.ts
