+++ apps/tradinggoose/lib/copilot/components/user-input/user-input.tsx
