+++ apps/tradinggoose/lib/copilot/components/user-input/workspace-entity-mentions.ts
