+++ apps/tradinggoose/lib/copilot/components/welcome/welcome.test.tsx
