+++ apps/tradinggoose/lib/copilot/components/welcome/welcome.tsx
