+++ apps/tradinggoose/lib/copilot/live-contexts.test.ts
