+++ apps/tradinggoose/lib/copilot/live-contexts.ts
