+++ apps/tradinggoose/lib/copilot/workspace-entities.test.ts
