+++ apps/tradinggoose/lib/copilot/workspace-entities.ts
